Improve batch point update tests, add multivec, extract into module
